+++ src/main/java/edu/ie3/simona/api/simulation/ExtCoSimulation.java
@@ -105,5 +105,4 @@ public abstract class ExtCoSimulation extends ExtSimulation {
       return new ExtResultDataConnection(resultEntities);
     }
   }
-
 }